requirements.txt
Flask==3.0.2
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.1.1
flask-bootstrap==3.3.7.1
pyyaml==6.0.1
flask-cors==4.0.0